TK3-67: Merge feature/TK3-67_t3d-deepl-translator-php-bugfixes to develop

See merge request tuk/tuk-projekte/tuk-typo3-extension/tuk-autotranslate!32

+++ Classes/UserFunction/FormEngine/AutotranslateLanguagesItems.php
@@ -44,16 +44,16 @@ class AutotranslateLanguagesItems {
             return;
         }
 
-        $sitePid = $row['pid'];
+        $sitePid = (int)$row['pid'];
         // pid < 0 respects the record (content or page) to insert after
         // pid > 0 and numeric represent the page uid to insert on
         if (is_numeric($row['pid']) && $row['pid']<0) {
-            $sitePid = Records::getRecord($table, abs($row['pid']), 'pid');
+            $sitePid = (int)Records::getRecord($table, abs($row['pid']), 'pid');
         }
 
         // set site pid on editing pages
         if ($row['pid'] == 0 && $table == 'pages' && is_numeric($row['uid'])) {
-            $sitePid = $row['uid'];
+            $sitePid = (int)$row['uid'];
         }
 
         // step out on create pages on root level
